Update Logitech Media Server to 8.5.1
Changelog is available here:

https://lyrion.org/getting-started/changelog-lms8/#version-851-2024-04-15-3b46196aa

The biggest change is a "Report Analytics Data" plugin that collects information
about the OS, hardware, perl version, etc. (NOT music or what is being
played). This is collected for statistics for the community for
deciding what to support in the future.

More info:

https://forums.slimdevices.com/forum/user-forums/general-discussion/1697108-request-for-comments-collecting-usage-data

https://lyrion.org/analytics/learn-more/

https://lyrion.org/analytics/

The plugin starts after some minutes and runs weekly. It can be disabled, but user are
encouraged to use it.