Add a port for Lyrion Music Server (formerly Logitech Media Server)
With version 9, the project gets a new name. The basics are mostly the
same.

In addition, after discussion with the Lyrion community, the firmware
and other files that were not GPL-licensed have been removed. This
means that binary packages can be distributed by the FreeBSD project
instead of forcing a local build. In addition, I have checked that
this also works for aarch64.

Also made changes so that Lyrion uses some of the Perl 5 ports instead
of the vendor versions. For now this is p5-Audio-Scan, p5-EV, and
p5-Image-Scale, p5-Encode-Detect, p5-Data-Dump, p5-Compress-Raw-Zip,
and p5-DBD-SQLite. This should make things compile with Perl 5.42 as
well (and reduce this port's build time as well).

As maintainer of logitechmediaserver, I am removing the
audio/logitechmediaserver port in the same commit. With Perl 5.40+,
this works without changes, whereas audio/logitechmediaserver needs
lots of messaging. So, it's time to do the upgrade. I am also removing the
other plugins as they are unmaintained and I'm not sure the versions
in the port tree worked with the version og logitechmediaserver in the
ports tree.

Changelog is available at:

https://lyrion.org/getting-started/changelog-lms9/#version-903-2025-10-03-9977737c1