Add a port for Lyrion Music Server (formerly Logitech Media Server)
With version 9, the project gets a new name. The basics are mostly the
same.

In addition, after discussion with the Lyrion community, the firmware
and other files that were not GPL-licensed have been removed. This
means that binary packages can be distributed by the FreeBSD project
instead of forcing a local build. In addition, I have checked that
this also works for aarch64.

As maintainer of logitechmediaserver, I am removing the
audio/logitechmediaserver port in the same commit. With Perl 5.40,
this works without changes, whereas audio/logitechmediaserver needs
messaging. So, it's time to do the upgrade. I am also removing the
other plugins as they are unmaintained and I'm not sure the versions
in the port tree worked with the version og logitechmediaserver in the
ports tree.

Changelog is available at:

https://lyrion.org/getting-started/changelog-lms9/#version-902-2025-03-13-1470c9412