Add a port for Lyrion Music Server (formerly Logitech Media Server)
With version 9, the project gets a new name. The basics are mostly the
same.

In addition, after discussion with the Lyrion community, the firmware
and other files that were not GPL-licensed have been removed. This
means that binary packages can be distributed by the FreeBSD project
instead of forcing a local build.

As maintainer of logitechmediaserver I am keeping the 8.x port around
until there is at least a minor update to Lyrion, even though 9.0 seems
pretty solid to start with.

Changelog is available at:

https://lyrion.org/getting-started/changelog-lms9/#version-901-2025-01-09-e3effbe91